+++ lib/main.dart
@@ -1,9 +1,18 @@
 import 'package:flutter/material.dart';
-import 'home_page.dart';
-import 'search_page.dart';
+import 'dart:math' as math;
 
 void main() {
   runApp(const MyApp());
+}
+
+class Contact {
+  String image;
+  String name;
+  String mobileNumber;
+  DateTime date;
+  bool isIncoming;
+
+  Contact(this.image, this.name, this.mobileNumber, this.date, this.isIncoming);
 }
 
 class MyApp extends StatelessWidget {
@@ -12,107 +21,238 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Flutter Demo 2',
       theme: ThemeData(
-        iconTheme: IconThemeData(color: Color.fromARGB(40,40, 40, 1)),
-        appBarTheme: AppBarTheme(elevation: 1,
-        color: Colors.white,
-         iconTheme: IconThemeData(color: Color.fromARGB(40, 40, 40, 1))), 
-        primarySwatch: Colors.pink,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(),
+      debugShowCheckedModeBanner: false,
+      home: const MyHomePage(title: 'Contacts App'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
   @override
   State<MyHomePage> createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int currentPage = 0;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: currentPage == 1 ? SearchPage() : HomePage(),
-      bottomNavigationBar: BottomAppBar(
+  int _selectedIndex = 2;
+  static const TextStyle optionStyle =
+      TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
+  static late List<Widget> _pages;
+
+  _MyHomePageState() {
+    _pages = [
+      buildContactsList(),
+      buildFavoritesGridView(),
+      // Text('hello'),
+      Text(
+        'Index 2: School',
+        style: optionStyle,
+      ),
+    ];
+  }
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  var contacts = [
+    Contact(
+      'https://i.pravatar.cc/300',
+      'Ahmed',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(seconds: 3),
+      ),
+      true,
+    ),
+    Contact(
+      'https://i.pravatar.cc/301',
+      'Ali',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 1),
+      ),
+      false,
+    ),
+    Contact(
+      'https://i.pravatar.cc/302',
+      'Kamal',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 3),
+      ),
+      true,
+    ),
+    Contact(
+      'https://i.pravatar.cc/303',
+      'Mohammad',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 5),
+      ),
+      true,
+    ),
+    Contact(
+      'https://i.pravatar.cc/304',
+      'Mohammad',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 5),
+      ),
+      false,
+    ),
+    Contact(
+      'https://i.pravatar.cc/305',
+      'Hussein',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 6),
+      ),
+      false,
+    ),
+    Contact(
+      'https://i.pravatar.cc/306',
+      'Aboud',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 7),
+      ),
+      false,
+    ),
+    Contact(
+      'https://i.pravatar.cc/307',
+      'Osama',
+      '71766137347',
+      DateTime.now().add(
+        const Duration(days: 6),
+      ),
+      false,
+    ),
+  ];
+
+  Widget buildFavoritesGridView() {
+    return Column(
+      children: [
+        Text('Favorites'),
+        Divider(
+          thickness: 4,
+        ),
+        Expanded(
+          child: GridView.count(
+            crossAxisCount: 3,
+            children: List.generate(5, (index) {
+              var personColor =
+                  Color((math.Random().nextDouble() * 0xFFFFFF).toInt())
+                      .withOpacity(1.0);
+              return Center(
+                child: Container(
+                  width: 120,
+                  height: 120,
+                  child: Text(
+                    contacts[index].name[0],
+                    style: TextStyle(fontSize: 40),
+                  ),
+                  alignment: Alignment.center,
+                  decoration:
+                      BoxDecoration(shape: BoxShape.circle, color: personColor),
+                ),
+              );
+            }),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget buildContactItem(Contact _contact) {
+    return Card(
+      child: Padding(
+        padding: const EdgeInsets.all(8.0),
         child: Row(
           children: [
-            IconButton(
-              icon: Icon(
-                Icons.home,
-                color: currentPage == 0
-                    ? Color.fromARGB(203, 73, 101, 1)
-                    : Color.fromARGB(40, 40, 40, 1),
-              ),
-              onPressed: () {
-                setState(() {
-                  currentPage = 0;
-                });
-              },
+            CircleAvatar(
+              backgroundImage: NetworkImage(_contact.image),
             ),
-            Spacer(),
-            IconButton(
-              icon: Icon(
-                Icons.search,
-                color: currentPage == 1
-                    ? Color.fromARGB(203, 73, 101, 1)
-                    : Color.fromARGB(40, 40, 40, 1),
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    _contact.name,
+                    style: const TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  Text(_contact.mobileNumber),
+                ],
               ),
-              onPressed: () {
-                setState(() {
-                  currentPage = 1;
-                });
-              },
             ),
-            Spacer(),
-            IconButton(
-              icon: Icon(
-                Icons.ondemand_video,
-                color: currentPage == 2
-                    ? Color.fromARGB(203, 73, 101, 1)
-                    : Color.fromARGB(40, 40, 40, 1),
-              ),
-              onPressed: () {
-                setState(() {
-                  currentPage = 2;
-                });
-              },
+            Text(_contact.date.toIso8601String().split('T').first),
+            Expanded(
+              child: Container(),
             ),
-            Spacer(),
-            IconButton(
-              icon: Icon(
-                Icons.card_travel,
-                color: currentPage == 3
-                    ? Color.fromARGB(203, 73, 101, 1)
-                    : Color.fromARGB(40, 40, 40, 1),
-              ),
-              onPressed: () {
-                setState(() {
-                  currentPage = 3;
-                });
-              },
-            ),
-            Spacer(),
-            IconButton(
-              icon: Icon(
-                Icons.person,
-                color: currentPage == 4
-                    ? Color.fromARGB(203, 73, 101, 1)
-                    : Color.fromARGB(40, 40, 40, 1),
-              ),
-              onPressed: () {
-                setState(() {
-                  currentPage = 4;
-                });
-              },
-            ),
+            if (_contact.isIncoming)
+              Icon(
+                Icons.arrow_downward,
+                color: Colors.red,
+              )
+            else
+              Icon(
+                Icons.arrow_upward,
+                color: Colors.green,
+              )
           ],
         ),
       ),
     );
   }
-}
 
+  Widget buildContactsList() {
+    return ListView.builder(
+      itemBuilder: (_context, index) {
+        return buildContactItem(contacts[index]);
+      },
+      itemCount: contacts.length,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: _pages[_selectedIndex],
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'Recent',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.favorite),
+            label: 'Favorites',
+          ),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.access_time_outlined),
+              label: 'School',
+              activeIcon: Icon(Icons.access_time_filled)),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: Colors.amber[800],
+        onTap: _onItemTapped,
+      ),
+    );
+  }
+}
